refactor: update markdownKatex import to use markdownHelper
